Deleted Wheel Flux and fixed esc indicator
