+++ includes/TimeUtils.h
@@ -1,0 +1,9 @@
+#ifndef RENNSSELAERMOTORSPORT_TIME_UTILS_H
+#define RENNSSELAERMOTORSPORT_TIME_UTILS_H
+
+#include <time.h>
+
+double getCurrentTime();
+double measureInterval(double lastIntervalTime);
+
+#endif
