+++ includes/WheelLocation.h
@@ -1,0 +1,11 @@
+#ifndef RENNSSELAERMOTORSPORT_WHEEL_LOCATION_H
+#define RENNSSELAERMOTORSPORT_WHEEL_LOCATION_H
+
+typedef enum {
+    front_left,
+    front_right,
+    back_left,
+    back_right
+} WHEEL_LOCATION;
+
+#endif
